Call resetConnection() from within CS104_Connection_create()
CS104_Connection_connect() relies on the value of self->running and
self->failure in order to know when to return. However, these values are set
only after the newly created thread enters handleConnection(), and are not
initialised by CS104_Connection_create(). This means that
CS104_Connection_connect() might inspect the value of self->running or
self->failure before the thread gets a chance to run, thus accessing
uninitialised memory and potentially returning early, before a connection is
actually made.

This commit fixes this, as well as a potential issue where the connection would
be stopped, connection parameters changed, and then restarted. In the old
implementation this would lead to an out-of-bounds access if the 'k' parameter
was larger than before, since the sentASDUs array was not allocated to take the
new parameter value into account. Since sentASDUs is initialised to NULL inside
_create(), it is possible to safely call free() from within resetConnection()
as well as _destroy().