Revert (fd == -1) check
This was added in attempts to fix a double-free
error. While it did seem to resolve the problem,
the solution itself is not ideal, and could be
responsible for a memory leak.

Tests after removing the code did not seem to
cause double-free errors. This could be because
the root issue was fixed upstream. More testing
is necessary to confirm.